Rename project pop
