--- src/plugin_output_processing/providers/__init__.py
@@ -1,4 +1,0 @@
-from plugin_output_processing.providers.openai import OpenAI
-from plugin_output_processing.providers.ollama import Ollama
-
-__all__ = ["OpenAI", "Ollama"]
