+++ src/pop/globals.py
