+++ src/pop/providers/openai.py
@@ -18,8 +18,8 @@
 import openai
 
 
-from plugin_output_processing.globals import Provider
-from plugin_output_processing.providers.base import BaseProvider
+from pop.globals import Provider
+from pop.providers.base import BaseProvider
 
 
 class OpenAI(BaseProvider):
